rewrote auto speed selection code
The new code distinguishes between minor, normal and major reception errors.
Major errors trigger an immediate speed switch, but in case of normal errors
a speed switch is only allowed each second command frame.

This acts as a rate-limiter on speed switches and gives UARTs (like the
Moschip 9835) more time to sync to the baudrate.