dir2atr: extend error reporting and set exit code to 1 on errors
Errors when adding files, for example when the files didn't fit
into the image, only caused warning messages to be printed.

Change this to error messages and set the exit code to 1 in case
of errors so using dir2atr in automated build systems can properly
fail the build in case of issues.

Signed-off-by: Matthias Reichl <hias@horus.com>